Changed locations of functions and removed files.
- No longer need checkout.go, put.go, or remove.go.
- Moved functions to where they should be.
