--- simple-rest-server/put.go
@@ -1,1 +1,0 @@
-package main
